+++ go/action_kit_api/action_kit_api_test.go
@@ -87,6 +87,10 @@ func TestActionDescription(t *testing.T) {
 		}),
 		Kind:        Attack,
 		TimeControl: Internal,
+		Hint: Ptr(ActionHint{
+			Type:    HintInfo,
+			Content: "foobar",
+		}),
 		Parameters: []ActionParameter{
 			{
 				Label:        "Wait for rollout completion",